Refactored synthetic data losses and metrics to model.py
